restructured files and project structure related to the Objective-C port
